+++ src/context/applets/photos/PhotosApplet.h
@@ -54,7 +54,7 @@ class PhotosApplet : public Context::Applet, public EngineObserver
 
         // inherited from EngineObserver
         virtual void engineNewTrackPlaying();
-        virtual void enginePlaybackEnded( int finalPosition, int trackLength, PlaybackEndedReason reason );
+        virtual void enginePlaybackEnded( qint64 finalPosition, qint64 trackLength, PlaybackEndedReason reason );
         
     public slots:
         void    dataUpdated( const QString& name, const Plasma::DataEngine::Data& data );
