+++ src/widgets/ProgressWidget.h
@@ -49,7 +49,7 @@ class ProgressWidget : public QWidget, public EngineObserver
         virtual void engineStateChanged( Phonon::State state, Phonon::State oldState = Phonon::StoppedState );
         virtual void engineTrackLengthChanged( qint64 milliseconds );
         virtual void engineNewTrackPlaying();
-        virtual void enginePlaybackEnded( int finalPosition, int trackLength, PlaybackEndedReason reason );
+        virtual void enginePlaybackEnded( qint64 finalPosition, qint64 trackLength, PlaybackEndedReason reason );
 
     private:
         TimeLabel *m_timeLabelLeft;
